# docs/api/map.openapi.yaml
openapi: 3.1.0
info:
  title: KApp Map API
  version: 3.6.0
  summary: Schematic campus map and space search for Fundacion Universitaria Konrad Lorenz.
  description: |
    Digital campus map for Fundacion Universitaria Konrad Lorenz (Bogota, Colombia).

    ## What this service is (and is not)

    This map is **not** a geospatial service. There is no basemap, no GPS and no external map
    provider, and a floor is never positioned by coordinates: it is a drawing, described as data
    and drawn by the client, not a set of floor plan images either. What version 3.1 adds is the
    ground a campus stands on - see below.

    ## Version 2: what the campus survey changed

    Version 1 was designed before anybody walked the campus. The survey of the real buildings
    (`docs/map/LEVANTAMIENTO.md`) broke five of its assumptions, and version 2 fixes each:

    | Found on campus | Version 1 | Version 2 |
    | --- | --- | --- |
    | A mezzanine between the first and second floors; an entrance at `P0`; a terrace numbered 5 | Floors addressed by an integer `level` | Floors addressed by a **`code`** (`S1`, `P0`, `P1`, `MEZZ`, `T`); `level` is a decimal that only orders them |
    | Bienestar's wings are west and east, and its doors carry no wing mark | A closed `NORTE`/`SUR`/`CENTRAL` enum, guessed from a `-N`/`-S`/`-C` suffix | **Wings declared per building**, each with the suffix its doors carry, if any |
...

...
    photographs of posted plans) or `VERIFIED` (walked).

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
  /api/map/buildings:

    get:
      x-roles: [ROLE_GUEST, ROLE_STUDENT, ROLE_PROFESSOR, ROLE_STAFF, ROLE_ADMIN]
      operationId: listBuildings
      summary: List buildings
      description: |
        Every building with its wings and floors, ordered by code. `q` matches the code, the name
        and every alias, ignoring case and accents.

        Allowed roles: `ROLE_GUEST`, `ROLE_STUDENT`, `ROLE_PROFESSOR`, `ROLE_STAFF`, `ROLE_ADMIN`.
      tags: [Buildings]
      parameters:
        - $ref: '#/components/parameters/CampusQuery'
        - name: q
          in: query
          required: false
          description: Code, name or alias, ignoring case and accents.
          schema: { type: string, minLength: 1, maxLength: 120 }
          example: psicologia
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: The buildings, ordered by code.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Building' }
              examples:
                Buildings: { $ref: '#/components/examples/BuildingList' }
        '304': { $ref: '#/components/responses/NotModified' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }

    post:
      x-roles: [ROLE_ADMIN]
      operationId: createBuilding
      summary: Create a building
      description: |
        Registers a new building with its wings and floors. The `code` must be unique across the
        map; floor codes and levels, and wing codes, must be unique within the building.

        Allowed roles: `ROLE_ADMIN` only.
      tags: [Buildings]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BuildingRequest' }
            example:
              code: BI
              name: Bienestar Institucional
              campus: Sede Principal
              aliases: [Bienestar, Edificio de bienestar]
              wings:
                - { code: OCC, name: Ala occidental, note: 'Bienestar: juegos, gimnasio, terraza.' }
                - { code: OR, name: Ala oriental, note: Salones y laboratorios. Se cruza por el P1 o la terraza. }
              floors:
                - { code: P1, level: 1, name: Piso 1, width: 640, height: 480 }
                - { code: T, level: 5, name: Terraza, width: 640, height: 480 }
      responses:
        '201':
          description: The building was created.
          headers:
            Location:
              description: Path of the created building.
              schema: { type: string }
              example: /api/map/buildings/BI
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Building' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/map/buildings/{code}:

    parameters:
      - $ref: '#/components/parameters/BuildingCodePath'

    get:
      x-roles: [ROLE_GUEST, ROLE_STUDENT, ROLE_PROFESSOR, ROLE_STAFF, ROLE_ADMIN]
      operationId: getBuildingByCode
      summary: Get a building by code
      description: |
        Allowed roles: `ROLE_GUEST`, `ROLE_STUDENT`, `ROLE_PROFESSOR`, `ROLE_STAFF`, `ROLE_ADMIN`.
      tags: [Buildings]
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: The building, its wings and its floors in vertical order.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Building' }
              examples:
                BloqueB: { $ref: '#/components/examples/BloqueB' }
        '304': { $ref: '#/components/responses/NotModified' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }

    put:
      x-roles: [ROLE_ADMIN]
      operationId: updateBuilding
      summary: Replace a building
      description: |
        Replaces the building, wings and floors included: a floor or a wing omitted here is
        deleted, and one that still has spaces is refused with `409`. Each floor keeps its
        `version`, which belongs to the layout save. A changed code, campus, floor level or wing
        suffix is copied onto every space of the building.

        Allowed roles: `ROLE_ADMIN` only.
      tags: [Buildings]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/BuildingRequest' }
      responses:
        '200':
          description: The building as stored.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Building' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }

    delete:
      x-roles: [ROLE_ADMIN]
      operationId: deleteBuilding
      summary: Delete a building
      description: |
        The building must be empty: one that still has spaces is refused with `409` and nothing
        is deleted. Never cascades.

        Allowed roles: `ROLE_ADMIN` only.
      tags: [Buildings]
      responses:
        '204': { description: Deleted. }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/map/buildings/{code}/floors/{floorCode}:

    parameters:
      - $ref: '#/components/parameters/BuildingCodePath'
      - $ref: '#/components/parameters/FloorCodePath'

    get:
      x-roles: [ROLE_GUEST, ROLE_STUDENT, ROLE_PROFESSOR, ROLE_STAFF, ROLE_ADMIN]
      operationId: getFloor
      summary: Get a floor and every space on it
      description: |
        The call that draws a floor: a drawing `width` x `height` units, origin at the top left,
        with the building's `outline`, each space as the polygon in its `shape` with its `doors`
        on it, and the corridors along their paths. Spaces without `shape` are on the floor but not
        drawn yet - list them beside the drawing rather than dropping them. The building's wings
        come along so each space's `wing` code can be shown by name.

        Allowed roles: `ROLE_GUEST`, `ROLE_STUDENT`, `ROLE_PROFESSOR`, `ROLE_STAFF`, `ROLE_ADMIN`.
      tags: [Buildings]
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: The floor and its spaces, ordered by code.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FloorDetail' }
              examples:
                Mezzanine: { $ref: '#/components/examples/MezzanineFloor' }
        '304': { $ref: '#/components/responses/NotModified' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /api/map/buildings/{code}/floors/{floorCode}/layout:

    parameters:
      - $ref: '#/components/parameters/BuildingCodePath'
      - $ref: '#/components/parameters/FloorCodePath'

    put:
      x-roles: [ROLE_ADMIN]
      operationId: saveFloorLayout
      summary: Save everything drawn on a floor at once
      description: |
        What the floor editor sends after someone walks a floor. `spaces` is the **complete**
        list: a space of this floor missing from it is deleted, a new one is created, the rest are
        replaced. **All or nothing** - if any space is wrong, nothing is written and every problem
        is reported against its position (`spaces[3].shape`).

        `version` is the floor's version as the editor loaded it. When someone saved the floor
        since, the save is refused with `409` rather than silently replacing their work; reload
        and reapply. A successful save returns the floor with its next version.

        Also `409`: removing a space that rooms on another floor name as their `accessVia`.

        Allowed roles: `ROLE_ADMIN` only.
      tags: [Buildings]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/FloorLayoutRequest' }
            example:
              version: 0
              width: 400
              height: 240
              status: VERIFIED
              accessibility: STAIRS_ONLY
              note: Solo por la escalera, desde la recepción.
              corridors: []
              spaces:
                - code: B-MEZZ-01
                  name: Laboratorio de Interactividad
                  typeCode: LAB
                  aliases: [interactividad]
                  shape: [{ x: 80, y: 80 }, { x: 240, y: 80 }, { x: 240, y: 160 }, { x: 80, y: 160 }]
                  doors: [{ from: { x: 140, y: 160 }, to: { x: 160, y: 160 } }]
                  accessVia: B-ESC
                - code: B-MEZZ-02
                  name: Sala MAC
                  typeCode: COMPUTER_LAB
      responses:
        '200':
          description: The floor as saved, with its next version.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FloorDetail' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Accessibility:
      type: string
      description: |
        Whether a floor or a space is reachable without stairs. `UNKNOWN` is the default and
        **must not be shown as either answer** - most of the campus has not been checked yet.
      enum: [STEP_FREE, STAIRS_ONLY, UNKNOWN]
    FloorStatus:
      type: string
      description: |
        `UNMAPPED` - the floor exists, nothing is drawn. `DRAFT` - drawn from photographs of posted
        plans, right in shape, unconfirmed in use. `VERIFIED` - walked and corrected on site.
      enum: [UNMAPPED, DRAFT, VERIFIED]
...
    Compass:
      type: string
      enum: [NORTH, EAST, SOUTH, WEST]
      description: |
        A direction on the ground. A floor's `top` says which one its drawing's top edge faces, so a
        client can turn the floor north up or draw a compass that points the right way. Turning a
        drawing a quarter turn clockwise brings what was on its left to the top.
      example: EAST
...
    Placement:
      type: object
      description: |
        Where a building's drawing lies on the ground: its top-left corner, the bearing its top
        edge faces and how long one unit of it is. Its floors' `top` is the same bearing to the
        nearest quarter turn.
      required: [origin, bearing, metresPerUnit]
      properties:
        origin: { $ref: '#/components/schemas/GeoPoint' }
        bearing:
          type: number
          minimum: 0
          exclusiveMaximum: 360
          description: Degrees clockwise from true north that the drawing's top edge faces.
          example: 127.8
        metresPerUnit:
          type: number
          exclusiveMinimum: 0
          description: How long one unit of the drawing is on the ground, in metres.
          example: 0.02951
    FootprintPart:
      type: object
      description: |
        One part of a building from above, as the city's cadastre records it or, where the
        cadastre falls short, as it was found on site. A building is seldom one block - the
        Edificio Central is a strip of five floors along the Calle 63, a core of eight and a south
        wing over its auditorium - and each part rises its own number of floors.
      required: [floors, basements, ring]
      properties:
        lot: { type: string, maxLength: 30, description: "The cadastral lot the part stands on. Absent for a part the cadastre does not record, found on site.", example: '008213024019' }
        floors: { type: integer, minimum: 0, maximum: 200, description: "How many floors it rises above the street: the cadastre's count, unless checked on site.", example: 8 }
        lowestFloor: { type: integer, minimum: 1, maximum: 200, description: "The lowest floor above the street it takes in, when it is not the first: 2 for what the upper floors carry out over a portico or a sidewalk, which the cadastre, seeing from above, draws as if it reached the ground. Absent for a part that stands on the street.", example: 2 }
        basements: { type: integer, minimum: 0, maximum: 20, description: How many it goes below., example: 2 }
        wing: { type: string, maxLength: 10, description: "The code of the building's wing it belongs to, when known.", example: C }
        ring: { $ref: '#/components/schemas/Ring' }
...
    Ring:
      type: array
      description: The outline of an area, closed - its first point repeated at the end.
      minItems: 4
      items: { $ref: '#/components/schemas/Coordinate' }
...
    Wing:
      type: object
      description: One arm of a building, as that building names it.
      required: [code, name]
      properties:
        code: { type: string, pattern: '^[A-Z0-9]{1,8}$', example: S }
        name: { type: string, maxLength: 60, example: Ala sur }
        doorSuffix:
          type: string
          maxLength: 4
          description: What the doors in this wing append to the room number. Absent when they append nothing.
          example: -S
        note:
          type: string
          maxLength: 300
          description: How to get into or across this wing, where that is not obvious.
          example: Se cruza por el P1 o por la terraza.
    Floor:
      type: object
      description: One floor of a building, described as a drawing the client scales and draws.
      required: [code, level, name, width, height]
      properties:
        code:
          type: string
          pattern: '^[A-Z0-9]{1,8}$'
          description: Identifier within the building, and the path segment the API uses. `S1` basement, `P0`, `P1`, `MEZZ`, `T` terrace.
          example: MEZZ
        level:
          type: number
          minimum: -5
          maximum: 99
          description: Vertical order only. Decimal so a mezzanine sits between its floors.
          example: 1.5
        name: { type: string, maxLength: 60, example: Mezzanine }
        status:
          $ref: '#/components/schemas/FloorStatus'
        accessibility:
          $ref: '#/components/schemas/Accessibility'
        note:
          type: string
          maxLength: 300
          description: How to get here when it is not obvious.
          example: Solo por la escalera, desde la recepción.
        width:
          type: integer
          minimum: 1
          maximum: 20000
          description: The drawing's width, in its own units. Every point on the floor lies within it.
          example: 1640
        height: { type: integer, minimum: 1, maximum: 20000, example: 590 }
        top:
          $ref: '#/components/schemas/Compass'
          description: The direction on the ground the drawing's top edge faces. Absent until somebody says - the posted plans are not drawn north up.
        outline:
          type: array
          maxItems: 400
          description: The building's walls around this floor, corners in order. Empty until traced.
          items: { $ref: '#/components/schemas/Point' }
        corridors:
          type: array
          items: { $ref: '#/components/schemas/Corridor' }
        version:
          type: integer
          format: int64
          readOnly: true
          description: Bumped by every layout save. Send it back with the next one; ignored when a whole building is written.
          example: 3
    Building:
      type: object
      required: [id, code, name, campus, aliases, wings, floors]
      properties:
        id: { type: string, example: 7c4b0a18-5e93-4a6f-8b21-9d0c5e7f1a34 }
        code: { type: string, maxLength: 10, example: B }
        name: { type: string, maxLength: 120, example: Bloque B }
        campus: { type: string, maxLength: 120, example: Sede Principal }
        description: { type: string, maxLength: 500 }
        aliases:
          type: array
          description: Other names people use for the building.
          items: { type: string }
        wings:
          type: array
          items: { $ref: '#/components/schemas/Wing' }
        floors:
          type: array
          description: Ordered by ascending level.
          items: { $ref: '#/components/schemas/Floor' }
        placement:
          $ref: '#/components/schemas/Placement'
          description: Where the building's drawing lies on the ground. Absent until somebody lays it.
        footprint:
          type: array
          description: The building from above, part by part, as the cadastre records it. Empty until somebody takes it from there.
          items: { $ref: '#/components/schemas/FootprintPart' }
        address: { type: string, maxLength: 200, description: "Its street address as people write it. Absent when nobody has said.", example: 'Cra. 9 Bis # 62-43' }
        sinuSedes:
          type: array
          description: The names SINU gives this building as a sede, exactly as its timetable prints them. Empty until somebody maps them in the portal. Since 3.6.
          items: { type: string, maxLength: 120 }
          example: [Sede Principal]
...
    BuildingRequest:
      type: object
      required: [code, name, campus, floors]
      properties:
        code: { type: string, minLength: 1, maxLength: 10 }
        name: { type: string, minLength: 1, maxLength: 120 }
        campus: { type: string, minLength: 1, maxLength: 120 }
        description: { type: string, maxLength: 500 }
        aliases:
          type: array
          items: { type: string, minLength: 1, maxLength: 120 }
        wings:
          type: array
          items: { $ref: '#/components/schemas/Wing' }
        floors:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/Floor' }
        placement:
          $ref: '#/components/schemas/Placement'
          description: Where the building's drawing lies on the ground. Omitted on `PUT`, the stored one is kept.
        footprint:
          type: array
          description: The building from above, part by part. Omitted on `PUT`, the stored one is kept; a part's `wing` must be one of the building's.
          items: { $ref: '#/components/schemas/FootprintPart' }
        address: { type: string, maxLength: 200, description: "Its street address as people write it. Omitted on PUT, the stored one is kept; empty clears it.", example: 'Cra. 9 Bis # 62-43' }
        sinuSedes:
          type: array
          description: The names SINU gives this building as a sede. Omitted on `PUT`, the stored ones are kept. A sede belongs to one building only; naming one another building has is a `409`. Since 3.6.
          items: { type: string, minLength: 1, maxLength: 120 }
...
    FloorDetail:
      type: object
      description: Everything needed to draw one floor in a single call.
      required: [code, level, name, status, accessibility, width, height, outline, corridors, version,
                 buildingId, buildingCode, buildingName, campus, wings, spaces]
      properties:
        code: { type: string, example: MEZZ }
        level: { type: number, example: 1.5 }
        name: { type: string, example: Mezzanine }
        status: { $ref: '#/components/schemas/FloorStatus' }
        accessibility: { $ref: '#/components/schemas/Accessibility' }
        note: { type: string }
        width: { type: integer }
        height: { type: integer }
        top: { $ref: '#/components/schemas/Compass' }
        outline:
          type: array
          items: { $ref: '#/components/schemas/Point' }
        corridors:
          type: array
          items: { $ref: '#/components/schemas/Corridor' }
        version: { type: integer, format: int64 }
        buildingId: { type: string }
        buildingCode: { type: string }
        buildingName: { type: string }
        campus: { type: string }
        wings:
          type: array
          items: { $ref: '#/components/schemas/Wing' }
        spaces:
          type: array
          description: Every space on the floor, placed or not, ordered by code.
          items: { $ref: '#/components/schemas/Space' }
...
    FloorLayoutRequest:
      type: object
      description: Everything drawn on one floor, saved at once.
      required: [version, width, height, spaces]
      properties:
        version:
          type: integer
          format: int64
          minimum: 0
          description: The floor's version as the editor loaded it.
        width: { type: integer, minimum: 1, maximum: 20000 }
        height: { type: integer, minimum: 1, maximum: 20000 }
        top:
          $ref: '#/components/schemas/Compass'
          description: Omit to keep the floor's.
        outline:
          type: array
          maxItems: 400
          items: { $ref: '#/components/schemas/Point' }
        status: { $ref: '#/components/schemas/FloorStatus' }
        accessibility: { $ref: '#/components/schemas/Accessibility' }
        note: { type: string, maxLength: 300 }
        corridors:
          type: array
          items: { $ref: '#/components/schemas/Corridor' }
        spaces:
          type: array
          description: The complete list. A space of this floor missing from it is deleted.
          maxItems: 500
          items: { $ref: '#/components/schemas/LayoutSpace' }
...
  headers:
    ETag:
      description: The version of this answer. Send it back in `If-None-Match` to be told when it changes. Since 3.6.
      schema: { type: string }
      example: '"3f2a9c14"'
  parameters:
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: The `ETag` of the answer the client already has. When nothing changed since, the answer is `304` with no body. Since 3.6.
      schema: { type: string }
      example: '"3f2a9c14"'
    CampusQuery:
      name: campus
      in: query
      required: false
      description: Restrict results to one campus (Sede), matched exactly.
      schema: { type: string, minLength: 1, maxLength: 120 }
      example: Sede Principal
    BuildingCodePath:
      name: code
      in: path
      required: true
      description: Building code.
      schema: { type: string, minLength: 1, maxLength: 10 }
      example: B
    FloorCodePath:
      name: floorCode
      in: path
      required: true
      description: Floor code within the building - `S1`, `P0`, `P1`, `MEZZ`, `T`.
      schema: { type: string, pattern: '^[A-Z0-9]{1,8}$' }
      example: MEZZ
...
  examples:
    BuildingList:
      summary: A search by alias - "psicologia" finds Bloque B
      value:
        - id: 7c4b0a18-5e93-4a6f-8b21-9d0c5e7f1a34
          code: B
          name: Bloque B
          campus: Sede Principal
          description: Psychology building. Research labs, library and student wellbeing services.
          aliases: [Bloque de psicología]
          wings: []
          floors:
            - { code: P1, level: 1, name: Piso 1, status: DRAFT, accessibility: STEP_FREE, width: 640, height: 440, outline: [], corridors: [], version: 0 }
            - { code: MEZZ, level: 1.5, name: Mezzanine, status: DRAFT, accessibility: STAIRS_ONLY, note: 'Solo por la escalera, desde la recepción.', width: 400, height: 240, outline: [], corridors: [], version: 0 }
            - { code: P2, level: 2, name: Piso 2, status: DRAFT, accessibility: STEP_FREE, width: 640, height: 440, outline: [], corridors: [], version: 0 }
            - { code: P3, level: 3, name: Piso 3, status: UNMAPPED, accessibility: UNKNOWN, width: 640, height: 440, outline: [], corridors: [], version: 0 }
    BloqueB:
      summary: A building with a mezzanine reachable only by stairs, and an alias
      value:
        id: 7c4b0a18-5e93-4a6f-8b21-9d0c5e7f1a34
        code: B
        name: Bloque B
        campus: Sede Principal
        description: Psychology building. Research labs, library and student wellbeing services.
        aliases: [Bloque de psicología]
        wings: []
        floors:
          - { code: P1, level: 1, name: Piso 1, status: DRAFT, accessibility: STEP_FREE, width: 640, height: 440, outline: [], corridors: [], version: 0 }
          - { code: MEZZ, level: 1.5, name: Mezzanine, status: DRAFT, accessibility: STAIRS_ONLY, note: 'Solo por la escalera, desde la recepción.', width: 400, height: 240, outline: [], corridors: [], version: 0 }
          - { code: P2, level: 2, name: Piso 2, status: DRAFT, accessibility: STEP_FREE, width: 640, height: 440, outline: [], corridors: [], version: 0 }
          - { code: P3, level: 3, name: Piso 3, status: UNMAPPED, accessibility: UNKNOWN, width: 640, height: 440, outline: [], corridors: [], version: 0 }
    MezzanineFloor:
      summary: A floor with a dependency that has nothing on its door
      value:
        code: MEZZ
        level: 1.5
        name: Mezzanine
        status: DRAFT
        accessibility: STAIRS_ONLY
        note: Solo por la escalera, desde la recepción.
        width: 400
        height: 240
        outline: []
        corridors: []
        version: 0
        buildingId: 7c4b0a18-5e93-4a6f-8b21-9d0c5e7f1a34
        buildingCode: B
        buildingName: Bloque B
        campus: Sede Principal
        wings: []
        spaces:
          - id: 5d1e9a33-0b6c-4f27-8e44-2a9c7b1d3f58
            code: B-MEZZ-01
            name: Laboratorio de Interactividad
            typeCode: LAB
            typeName: Laboratorio
            category: TEACHING
            buildingId: 7c4b0a18-5e93-4a6f-8b21-9d0c5e7f1a34
            buildingCode: B
            campus: Sede Principal
            floorCode: MEZZ
            floorLevel: 1.5
            aliases: [interactividad]
            shape: [{ x: 80, y: 80 }, { x: 240, y: 80 }, { x: 240, y: 160 }, { x: 80, y: 160 }]
            bounds: { x: 80, y: 80, width: 160, height: 80 }
            doors: [{ from: { x: 140, y: 160 }, to: { x: 160, y: 160 } }]
            accessVia: B-ESC
            effectiveAccessibility: STAIRS_ONLY
...
  responses:
    NotModified:
      description: Nothing changed since the `ETag` sent in `If-None-Match`. No body; keep using what you have.
    BadRequest:
      description: Validation failed
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Unauthorized:
      description: Missing or invalid token
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Forbidden:
      description: Authenticated but not allowed
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    NotFound:
      description: Resource not found
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }
    Conflict:
      description: Conflicts with current state
      content: { application/json: { schema: { $ref: '#/components/schemas/ApiError' } } }